moved some rotations3d functions into vector_utilities and updated tests.
