+++ halotools/utils/vector_utilities.py
@@ -6,8 +6,12 @@ from __future__ import (division, print_function, absolute_import,
                         unicode_literals)
 import numpy as np
 
-__all__=['elementwise_dot', 'elementwise_norm', 'normalized_vectors',
-         'angles_between_list_of_vectors']
+__all__=['elementwise_dot', 
+         'elementwise_norm', 
+         'normalized_vectors',
+         'angles_between_list_of_vectors', 
+         'vectors_normal_to_planes', 
+         'project_onto_plane']
 __author__ = ['Duncan Campbell', 'Andrew Hearin']
 
 
@@ -166,3 +170,77 @@ def angles_between_list_of_vectors(v0, v1, tol=1e-3, vn=None):
     return a   
 
 
+def vectors_normal_to_planes(x, y):
+    r""" 
+    Given a collection of 3d vectors x and y, return a collection of 
+    3d unit-vectors that are orthogonal to x and y.
+
+    Parameters
+    ----------
+    x : ndarray
+        Numpy array of shape (npts, 3) storing a collection of 3d vectors
+
+        Note that the normalization of `x` will be ignored.
+
+    y : ndarray
+        Numpy array of shape (npts, 3) storing a collection of 3d vectors
+
+        Note that the normalization of `y` will be ignored.
+
+    Returns
+    -------
+    z : ndarray
+        Numpy array of shape (npts, 3). Each 3d vector in z will be orthogonal
+        to the corresponding vector in x and y.
+
+    Examples
+    --------
+    Define a set of random 3D vectors
+
+    >>> npts = int(1e4)
+    >>> x = np.random.random((npts, 3))
+    >>> y = np.random.random((npts, 3))
+
+    now calculate a thrid set of vectors to a corresponding pair in `x` and `y`.
+
+    >>> normed_z = angles_between_list_of_vectors(x, y)
+    """
+    return normalized_vectors(np.cross(x, y))
+
+
+def project_onto_plane(x1, x2):
+    r"""
+    Given a collection of vectors, x1 and x2, project each vector
+    in x1 onto the plane normal to the corresponding vector x2.
+
+    Parameters
+    ----------
+    x1 : ndarray
+        Numpy array of shape (npts, 3) storing a collection of 3d points
+
+    x2 : ndarray
+        Numpy array of shape (npts, 3) storing a collection of 3d points
+
+    Returns
+    -------
+    result : ndarray
+        Numpy array of shape (npts, 3) storing a collection of 3d points
+
+    Examples
+    --------
+    Define a set of random 3D vectors.
+
+    >>> npts = int(1e4)
+    >>> x1 = np.random.random((npts, 3))
+    >>> x2 = np.random.random((npts, 3))
+
+    Find the projection of each vector in `x1` onto a plane defined by
+    each vector in `x2`.
+
+    >>> x3 = project_onto_plane(x1, x2)
+    """
+
+    n = normalized_vectors(x2)
+    d = elementwise_dot(x1,n)
+
+    return x1 - d[:,np.newaxis]*n
